+++ src/__tests__/app/nhs-app-template-submitted/page.test.tsx
@@ -2,7 +2,7 @@ import nav from 'next/navigation';
 import NhsAppTemplateSubmittedPage from '@app/nhs-app-template-submitted/[templateId]/page';
 import { render } from '@testing-library/react';
 import { getTemplate } from '@utils/form-actions';
-import { TemplateType } from '@utils/types';
+import { TemplateType } from '@utils/enum';
 
 jest.mock('@utils/form-actions', () => ({
   getTemplate: jest.fn().mockImplementation((templateId: string) => {
