+++ src/app/create-email-template/[templateId]/page.tsx
@@ -1,6 +1,7 @@
 import { CreateEmailTemplate } from '@forms/CreateEmailTemplate/CreateEmailTemplate';
-import { PageProps, TemplateType } from '@utils/types';
+import { PageProps } from '@utils/types';
 import { getTemplate } from '@utils/form-actions';
+import { TemplateType } from '@utils/enum';
 import { redirect, RedirectType } from 'next/navigation';
 
 const CreateEmailTemplatePage = async ({
