+++ src/app/create-text-message-template/[templateId]/page.tsx
@@ -1,7 +1,8 @@
 'use server';
 
 import { CreateSmsTemplate } from '@forms/CreateSmsTemplate/CreateSmsTemplate';
-import { PageProps, TemplateType } from '@utils/types';
+import { PageProps } from '@utils/types';
+import { TemplateType } from '@utils/enum';
 import { getTemplate } from '@utils/form-actions';
 import { redirect, RedirectType } from 'next/navigation';
