+++ src/components/forms/CreateSmsTemplate/server-action.ts
@@ -2,6 +2,7 @@ import { TemplateFormState } from '@utils/types';
 import { z } from 'zod';
 import { saveTemplate } from '@utils/form-actions';
 import { redirect, RedirectType } from 'next/navigation';
+import { MAX_SMS_CHARACTER_LENGTH } from '@utils/constants';
 
 const $CreateSmsTemplateSchema = z.object({
   smsTemplateName: z
@@ -10,7 +11,7 @@ const $CreateSmsTemplateSchema = z.object({
   smsTemplateMessage: z
     .string({ message: 'Enter a template message' })
     .min(1, { message: 'Enter a template message' })
-    .max(918, { message: 'Template message too long' }),
+    .max(MAX_SMS_CHARACTER_LENGTH, { message: 'Template message too long' }),
 });
 
 const $GoBackSchema = z.object({
