+++ src/components/molecules/TemplateNameGuidance/template-name-guidance.types.ts
@@ -1,4 +1,4 @@
-import { TemplateType } from '@utils/types';
+import { TemplateType } from '@utils/enum';
 
 export type TemplateNameGuidanceType = {
   template: keyof typeof TemplateType;
