+++ src/utils/enum.ts
@@ -1,0 +1,6 @@
+export enum TemplateType {
+  NHS_APP = 'NHS_APP',
+  SMS = 'SMS',
+  EMAIL = 'EMAIL',
+  LETTER = 'LETTER',
+}
